+++ apps/upgrade-v8-v9/1-before/public/manifest.json
